Refactor: Separate chatbot and quiz functions into gradio_app_chat.py and gradio_app_quiz.py, called by gradio_app.py
